fix: signInByTokenPage to signInByMagicLinkPage
